+++ test/mixin_bot/api/test_collectible.rb
@@ -88,7 +88,7 @@ module MixinBot
           media: {}
         }
       }
-      metahash = SHA3::Digest::SHA256.hexdigest [meta[:collection][:id], meta[:collection][:name], meta[:token][:id],
+      metahash = SHA3::Digest::SHA3_256.hexdigest [meta[:collection][:id], meta[:collection][:name], meta[:token][:id],
                                                  meta[:token][:name]].join
       memo = MixinBot.api.nft_memo collection, token_id, metahash